get_other_type works;  sort order box works; rescan works; filter is enabled; virtual fs setup dialog box is up; save setup; mkdir
link command is hossed.  It is more than hossed for some stoopid
reason, the thing is selecting the first file in the panel
just after I talked to it.  No matter how argumentative I get
with the code.  Have to fix next.

Commit the bits I forgot to commit yesterday.
-miguel.