Mon Mar 23 01:40:35 1998  Norbert Warmuth  <k3190@fh-sw.de>
* gnome/gtools.c (translate_gdk_keysym_to_curses): renamed KEY_BACKTAB
to KEY_BTAB (the name ncurses uses)

* gnome/Makefile.in: Install the icon-files and the layout-file from $srcdir.

* autogen.sh: added support for builddir != srcdir

* doc/Makefile.in: The manual pages are generated files and
located in the builddir and not in the srcdir.

* src/find.c (find_file): Use the same hotkey for the panelize button
as we use with the external panelize command.

* src/myslang.h: --with-ncurses didn't compile: renamed
KEY_BACKTAB to KEY_BTAB (the name ncurses uses)

* src/dlg.c (dlg_key_event): likewise

* src/tkmain.c (x_init_dlg): added dummy function to make it
compile

* vfs/Makefile.in, extfs/extfs.ini, extfs/hp48: New external
filesystem to view and copy files to/from the HP48.

Now it should be possible to build and install gnome edition with
builddir != srcdir.