Fixes for internationalization compile on some systems:
I have found the problem with all of those internationalization
problems that have been plagging us:

Using AM_whatever_GETTEXT Is not everything that is required:

we need to call this macro and the magic AC_LINK macro.  This macro
creates a link in intl from libgettext.h to libintl.h.  So, we should
be adding -I$builddir/intl and -L$builddir/lib options to the relevant
variables if USE_NLS is set to yes.

We need to propagate this to the gnome code.

Miguel.