Fix the ftp bug reported by Andrej Borsenkow. If we have to keep O_EXCL in retrieve_file_start2 then I will find a different fix.
Mon May  4 10:27:49 1998  Norbert Warmuth  <k3190@fh-sw.de>

* ftpfs.c (retrieve_file_start2): Don't create target file O_EXCL,
in copy_file_file we check existance of the target file and know
that we want to truncate it.


Mon May  4 10:21:31 1998  Norbert Warmuth  <k3190@fh-sw.de>

* file.c (copy_file_file): Schedule deletion of short target file
only when we created or truncated the target file and not already
when we decided to overwrite an existing file.