The first two patches fix Bug#1242. I will close that bug later.
Mon May 17 07:30:48 1999  Norbert Warmuth  <nwarmuth@privat.circular.de>

* configure.in: Added check for getpt ()

* src/subshell.c (pty_open_master): use getpt () if available to open
the master side of the pty. getpt () is a glibc extension and is
needed when glibc 2.1.x is used with a (linux) kernel without unix98
style ptys (if there is no /dev/ptmx getpt falls back to bsd style
ptys).

* src/filegui.c (check_progress_buttons): Don't update the Gpm mouse
cursor when we check for events. It causes a flickering cursor on a
different virtual console.