Fix the line-editing widget not to go completely crazy when seeing a literal newline (Shift+Enter or ^Q Enter).
The real cause of the problem is making no proper distinction between
length and width. Actually mbstrlen() returns the width (ignoring
non-printable characters), not the length. On the other hand, other pieces
of the code expect it to return the length. In fact, this function should be
called mbstrwidth() and we should have a completely different one called
mbstrlen(). A lot of problems regarding double-width characters still remain
due to this misconcept.

Another bug is counting non-printable characters (e.g. newline) as zero width.
The proper approach is to count them as 1 since we're printing a replacement
character for them.

Thanks to egmont who provided the patch

Signed-off-by: Patrick Winnertz <winnie@debian.org>