Ticket 38: return from Info panel to wrong directory.
When viewing two file lists, if user presses C-x i for info on
a panel's directory or file, the info pops up correctly in the
other panel. However, when the user closes the info screen, the
directory of the current panel is shown -- NOT the directory
which was being viewed before the info panel appeared.

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>