Ticket #1710: newlines are lost when pasting multiline text in editor.
When pasting multiline text with shift+mouse newline characters
are lost. Pasted text is not shown before any key is pressed.

Middle-mouse pasting in mc requires shift being pressed to keep mc
from interpreting the button press itself, so it is passing on to
xterm/gpm. That has the nice side effect that the pasted enter
characters are automatically shift-enter, which was simply interpreted
as newline-without-autoindent. However, the keybinding for shift-enter
have been lost in the course of #212 - it simply does nothing,
and consequently text is pasted without newlines.

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>