Ticket #1702: changed colors of alvays visible inactive menu bar.
If menu bar is always visible, default colors are used in menu.
If menu bar is visible when it is active, skin colors are used.

In this commit, only skin colors are applied to menu. Thanks to
Rashid N. Achilov for idea.
(see http://mail.gnome.org/archives/mc/2010-January/msg00087.html).

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>