Ticket #2091: Superfluous "Search done, Continue from begining?" dialog
When searching from the top of the file for something that's not there we get
"Search done, Continue from begining?" dialog before "Search string not found"
(and only if we choose "Yes" in the first one).

This should happen only if the last position>0 (either the place when search
began, or some match before).

Signed-off-by: Slava Zanko <slavazanko@gmail.com>