Ticket #2146: Misleading label "Status:" in the Info panel
The info panels shows useful information about the currently selected
file (atime, mtime, ctime for instance). However, the label "Status:"
for ctime is clearly misleading. As we don't have enough space to label
it "Status changed", I suggest to go for "Changed:" instead. I think
this will make it much clearer what this date is all about, as I have
already seen users suggesting that "Status" is a weird label for
"Created", which is obviously not the case.

Also, "Changed" label is consistent with "CHange time" that we use for
column headers in the list mode elsewhere.

Signed-off-by: Yury V. Zaytsev <yury@shurup.com>