Ticket #2237: Automatic date and version substitution for man pages
Currently the date and the version of Midnight Commander in the man
pages has to be changed manually every release. This is being done
sporadically and as the result, the man pages shipped for e.g. 4.7.2
release contain the 4.7.0-pre1 signature, which is quite confusing.

With this patch, the version is substituted with @DISTR_VERSION@ for all
the man pages  and the date is automatically updated to the English man
pages. Other man pages still have the manually updated date of the last
translation.

Signed-off-by: Yury V. Zaytsev <yury@shurup.com>