Ticket #2199: Buttonbar mouseclicks are off
Case 1: Make the terminal window 85 characters wide. Click (with the
mouse) on the right side of the "Copy" button on the bottom. See the
Move dialog appearing instead of the Copy dialog.

Case 2: Make the terminal 42 characters wide. Click on the bottom right
corner (the Re~ov button). MC invokes its menu or quits, depending on
the exact click location.

The patch that solves all these. The extra space is not assigned to the
first buttons, but distributed mostly evenly across buttons,
guaranteeing that the first five get exactly as much (or one character
less) than the second five, just as it works for the panels. Also, mouse
clicks use the exact same boundaries that were used when printing the
buttons.

Signed-off-by: Yury V. Zaytsev <yury@shurup.com>

Minor code cleanup, renamed local variable to avoid shadowing of global
variable, fixed formatting.

Signed-off-by: Slava Zanko <slavazanko@gmail.com>
Signed-off-by: Andrew Borodin <aborodin@vmail.ru>