Ticket #7: Show size of all files in directory (savannah)
Renamed the menu item "Show directory sizes" to "Show directory size(s)" to
emphasize its new meaning (in comparison with 4.6 branch).

Previously this menu item would compute the sizes of all directories
regardless of the selection and currently active item. Now by default it
computes the size of the currently active directory, unless there's a
selection or user is on "..".

Signed-off-by: Yury V. Zaytsev <yury@shurup.com>