Ticket #2333: MC hangs after exit
On some FreeBSD servers with tcsh shell the following bug occurs
sometimes: trying to exit from mc it hangs. More detail:
1. run mc, press F10, press Enter - hangs
2. run mc, press F10, choose Command->File->Exit - NOT hangs.

This hang up is a result of subshell_chdir() call after subshell
close (after exit_subshell() call). When the F10 key is handled in
dlg_try_hotkey() function, the do_select_widget() is called. As a
result, the WIDGET_FOCUS message is sent to the panel and panel
calls subshell_chdir(). But subshell is already closed at this
moment and midnight_dlg->running flag is reset.

The solution: don't send DLG_DRAW, DLG_FOCUS and DLG_UNFOCUS messages to
the inactive (with reset 'running' flag) dialog. Since dialog is
about destroying it is not needed in any visual changes.

Thanks Tempread for investigation of the problem root.

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>