Ticket #2344: Fix line jump when started as editor
Two common line jump protocols supported by mc are:

    * mcedit +lineno file
    * mcedit file:lineno

Unfortunately, in the implementation of the first protocol, if line
number equals to zero, which is often used to position the cursor at the
beginning of the file, it is completely ignored and file "+0" is opened
instead.

This patch fixes aforementioned issue.

Signed-off-by: Yury V. Zaytsev <yury@shurup.com>