Ticket #2395: scroll down the panel contents by mouse dragging.
Start mc, and make sure that at least one of the two displayed
directiories contains so many files that they cannot all be displayed
at once.

Click in this panel and hold the mouse button pressed, and then move
your mouse downwards. As you pass beyond the bottom of this list, it
should start scrolling.

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>