Ticket #2435: goto to the specified address in hex mode of viewer.
Viewer Goto dialog doesn't allow go to the specified offset. The value
of input line is rounded to the next line start. Such bevaviour
is useless in hex mode of viewer. In hex mode we must heve a
capability to go to the specified address as is without any round.

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>