Ticket #2294: mcview: incorrect starting offset for 'search again'
When you press F7 in mcview to repeat the previous search, it starts again from the
file offset that is 3 bytes after the beginning of the current match. This introduces
weird and clearly buggy symptoms. Couple of examples:
If the file contains

xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx

and you search for a single x character, and you keep on pressing F7,
every third character gets highlighted.

Signed-off-by: Slava Zanko <slavazanko@gmail.com>