Ticket #2201: file name length limit in tar archive
The problem: extracting deep files from an archive containing long directory/file
names, filenames are cut off at 100 characters (length of the whole path
inside the archive).

In general, the TarVFS is scheduled for total reimplementation. But this
fix is useful for stable releases.

Thanks AVFS (avf.sf.net) for the code of this fix.

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>