Ticket #2614: Editor word completion should ignore the current word
Currently the completion considers the word the cursor is on as a possible
completion, but this is quite annoying if the cursor is inside the word.
Such completion effectively inserts the rest of the word one more time, so
CamelCase becomes CamelCaseCase. If this is the only match, it completes
automatically, which is even worse. The current word shouldn't be used
for completion.

Signed-off-by: Slava Zanko <slavazanko@gmail.com>