Fixup update of command line after window resize.
The bug only occurs with subshell enabled.

When resizing the window while in the viewer, the subshell resizes
itself and hence prints its prompt again. This is captured and processed
by mc.

src/filemanager/layout.c:setup_panels() is executed and recalculates
properties of the panel, but for some reason this time mc_prompt does
contain all the invisible characters, they are not stripped off, hence
size calculation goes wrong.

Thanks Egmont Koblinger for the detailed description of the problem and
the idea of fix.

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>