Ticket #2278: Problem in the Copy operation
Previously, copying a file does not change access permissions of the destination
file if the check box "Preserve attributes" in the Copy window is not set.
For some time it is not so: if you uncheck the "Preserve attributes" then access
permissions of the destination file are set in accordance with the value of
umask, which can lead to information disclosure(i.e. security problem).

Signed-off-by: Slava Zanko <slavazanko@gmail.com>