Ticket #2783: directories is opened in the wrong panel.
mc dir1 dir2

The bug: order of parameters assignment depends on current_is_left
option in panels.ini file: if current_is_left=0, then dir1 is opened
in the left panel, dir2 in the right one. If current_is_left=1,
then dir1 is opened in the right panel, dir2 in the left one.

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>