Fix fish-shell prompt not working.
On 2.0beta+, the fish_prompt function will get overwritten after
starting the shell. Instead, create a custom function that gets
triggered after the prompt is displayed.

As a bonus, this also makes the users own prompt appear both when
the panel is visible and hidden.

This will conflict with the fish_prompt changes in
<https://www.midnight-commander.org/ticket/2742>, but I don't think
that should be a big issue.