Ticket #2164: clarify cursor position in mcviewer.
When viewing a file, the cursor's position is different in the slang
and ncurses builds.

With slang, the cursor is in the upper right corner, on the percent sign.

With ncurses, the cursor wraps around to the beginning of the next line,
that is, the top left corner of the actual file contents viewing area.

In this commit, cursor is moved to upper right corner unconditionally
after print of percentage value.

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>