Ticket #3097: fix recursive find file on Samba share.
src/filemanager/file.c (do_search): the algorithm executes mc_stat() on
a file (directory) and checks number of links. It assumes number of
links minus 2 is number of subdirectories available for recursive
search. While that assumption is valid on probably all Linux
filesystems, it is not valid on mounted Samba shares. For these, number
of links (as reported by stat or seen by 'ls -l') is not 1 for most
files and 2+ for directories, but zero for all files and directories
within the share (I guess that is how the kenrel samba module works
...). Therefore the optimization (that may limit the recursion) done by
subdirs_left breaks the recursive search.

Also this patch fixes recursive search on NFS shares provided by Windows
2012R2 NFS service (ticket #3202).

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>