Ticket #3244: add viewbold and viewselected missing from some skins.
Change "black" and "white" to "rgb000" and "rgb555" when it's rendered
in bold, because the legacy 16 colors change their actual color in
addition to going bold in most of the terminals which is undesired.
Second, there was some unintended difference between the two gray skins
in the file type colors.

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>