Ticket #3258: fix incorrect percentage in mcview hex mode.
The percent in the upper right corner normally corresponds to the
cursor. Except when the bottom of the file is displayed, then it's
100%.

If the file is taller than the window, you can walk downwards to its
end and the percent is always correct. Then walk back a screenful and
it stays at 100%, and suddenly jumps back by a lot when the file starts
scrolling.

If the file is shorter than the window, it's always at 100%.

mcview_calc_percent() is supposed to calculate the percentage at offset
p, yet it executes a special branch when "dpy_end == filesize".

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>