Sync with gnulib 2768ceb7994506e2cfba88be3b6bd13ef5440a90.
mountlist: don't use libmount to decide on dummy/remote.

Don't use the libmount routines to determine whether a file system is
dummy or remote, as they're not currently compatible. For example the
remoteness is determined on file system type (for which the list seems
incomplete), rather than simply checking for a ':' in the device name.
Also libmount currently determines that 'tmpfs' is a dummy file system
even though it has associated storage.

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>