Ticket #3259: mcview hex edit: UTF-8 chars are not updated.
UTF-8 characters show up nicely in the right column (followed by a space
to align with the number of bytes).

However, as soon as you edit the file (either by the hex codes, or by
moving to the right column and entering an UTF-8 character), that
character doesn't show up there.

Expected: during editing the file, the right hand side should
continuously be updated to look exactly as if that was the original
contents of the file.

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>