Ticket #3260: mcview hex edit: can't enter certain UTF-8 characters.
(mcview_handle_editkey): an is_printable() check was performed for every
byte of the UTF-8 which obviously doesn't make sense. This check
shouldn't be performed for 8-bit charsets either. If the given byte can
arrive as input, the file should be modified accordingly.
The TAB key shouldn't be accepted as literal, to allow moving back to
the hexview area.

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>