Ticket #3261: mcview hex edit: CJK overflow.
At certain terminal widths (including the default 80), if the last byte
of a line begins a CJK character then the corresponding glyps is drawn
in the last column, so it's replaced by a replacement symbol.

The default layout should, at all possible terminal widths, have an
extra last character column that is empty normally, but allows room for
a CJK to overflow here.

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>