Use PATH environment variable for Filtered View mc does not use the PATH environment variable to search for the executable when performing a Filtered View command.
The filtered view command is run with the mc_popen function in lib/utilunix.c. This function runs the process with the glib function g_spawn_async_with_pipes, which by default requires the full path of the executable.

This change adds the G_SPAWN_SEARCH_PATH flag to the g_spawn_async_with_pipes function call, which causes mc to use the PATH environment variable to search for the executable.