Ticket #3130: implement center scrolling of panel.
Behavior: cause the panel to begin scrolling when the cursor
reaches the middle of the panel, so that the cursor tends to stay in
the middle of the panel on long listings. Only when you reach the
beginning or the end of the listing will the cursor move to the first or
last file.

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>